backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
alembic==1.13.1
asyncpg==0.29.0  # PostgreSQL async driver
psycopg2-binary==2.9.9  # PostgreSQL sync driver

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Redis
redis==5.0.1
aioredis==2.0.1

# AI/ML Dependencies
langchain==0.1.0
langchain-community==0.0.10
langchain-openai==0.0.5
openai==1.6.1
chromadb==0.4.22
sentence-transformers==2.2.2
pypdf2==3.0.1
python-docx==1.1.0

# Data Processing
pandas==2.1.4
numpy==1.25.2
pydantic==2.5.2
pydantic-settings==2.1.0

# File Processing
python-magic==0.4.27
pillow==10.1.0
pytesseract==0.3.10

# Payment Integration
razorpay==1.4.0

# Utilities
python-dotenv==1.0.0
loguru==0.7.2
httpx==0.25.2
aiofiles==23.2.1

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.22.0
pytest-cov==4.1.0
black==23.12.1
flake8==6.1.0
mypy==1.8.0

# Monitoring
prometheus-client==0.19.0
